On Solaris, use /usr/ucb/install explicitly (we don't support their /usr/sbin/install) and install under /usr/lib instead of /lib.
Also on Solaris only, pass the "-e" option to the sub-make, which makes
a difference for Sun's make (e.g., DESTDIR would not be passed without
it) and is OK for GNU make as well.

Don't list the "all" target twice (which was done to make it the default
target).  Instead, list "default: all" as the first target (a target
named "default" that depends on "all").

CVS-ID: Makefile 1.30