Unexport _passwdqc_memzero in libpasswdqc
_passwdqc_memzero, as the underscore suggests, is an internal mechanism
not intended to be used outside passwdqc.

With this change, _passwdqc_memzero is no longer exported by
libpasswdqc, which means that libpasswdqc itself and all its users will
use their own copy of _passwdqc_memzero, which in turn introduces the
likelihood of link time optimization if the linker is able to figure out
that _passwdqc_memzero is never changed at runtime.

The change can be reconsidered later if this hypothetical link time
optimization will become more realistic.

CVS-ID: Makefile 1.33
CVS-ID: libpasswdqc.map 1.4
CVS-ID: passwdqc_check.c 1.34
CVS-ID: passwdqc_memzero.c 1.1