pam_passwdqc: enhance auto-generated policy descriptions
Fix grammar of descriptions containing unconditional
"%d characters" where a singular form is also possible.

Change wording from "that consists of %d characters" to
"containing at least %d characters" to make it clear
when the policy sets the lower limit.

Change the wording of the case when the policy have different
requirements for passwords containing characters from 3 and 4
classes to make it a single translatable sentence.

Wrap lines at a more consistent length.

Resolves: https://github.com/openwall/passwdqc/issues/3